[msbuild] Don't write actool/ibtool/etc outputs to FileWrites
MSBuild compares what is in FileWrites with the Outputs of the target
and any file that exists in FileWrites but *doesn't* exist in Outputs
gets deleted with the assumption that it is no longer needed via the
IncrementalClean MSBuild target.

Since these tasks cannot know what the outputs will be until the task
is run, we cannot use FileWrites.