[mmp] References are optional when not linking the app, so don't show errors in that case. Fixes #40786.
It is apparently common to have references to assemblies that can't
be resolved when building apps with the linker disabled, so we need
to support that.

So add a custom reference loading step that only shows a warning
for unresolvable assemblies.

Also bump maccore to add a test.

commit xamarin/maccore@190c67d8551def7bef9aa89c3100805ca57761ab
Author: Rolf Bjarne Kvinge <rolf@xamarin.com>
Date:   Tue May 10 18:14:04 2016 +0200

    [tests] Add test case for bug #40786.

    https://bugzilla.xamarin.com/show_bug.cgi?id=40786

https://bugzilla.xamarin.com/show_bug.cgi?id=40786