[gameplaykit] Replace QuadTreeWithMinPosition by managed code
Looks like it was an API added in 9.x but removed (in later betas?). The
introspection tests did find (in 9.x) the selector so it was not reported
as an error but it's not part of iOS 10 (so it fails now)

This removes the binding and replace it with a managed alternative
so there is no breaking change.

references:
[FAIL] GameplayKit.GKQuadTree : quadTreeWithMinPosition:maxPosition:minCellSize: