Build with path to the exact SDK version. Fixes #41597.
If no sdk version (-sdk_version) is passed to the native
linker, it tries to infer the SDK version from the
path to the -syslibroot argument.

In our case we use a versioned path to Xcode, but a general
symlink without the SDK version:

/Applications/Xcode73.app/Contents/Developer/Platforms/AppleTVOS.platform/Developer/SDKs/AppleTVOS.sdk

which means ld picked up the Xcode version as the SDK version,
since that's the first number part of the path [1], so we'd
end up with libraries whose SDK version was 73.

So instead use an SDK path with the SDK version, so that ld
finds the SDK version instead of the Xcode version.

[1] https://github.com/rolfbjarne/ld64/blob/266f4401b94ee73be4f4c18a0a220110c7fe71f0/src/ld/Options.cpp#L4005-L4020

https://bugzilla.xamarin.com/show_bug.cgi?id=41597