[watchps][gamekit] Enable GameKit on watchOS
Apple does not (yet?) ship the headers for GameKit on the watch simulator!

So we need to disable the default registrar.a for the platform since it
won't be possible to compile (static) until Apple fix this.

This also disable the AppleTV registrar.a as a similar issue exists with
ExternalAccesory on that platform