Modifying the way GAC elimination works so it will work in MSBuild
In MSBuild treating the property as a string will end up producing extra semicolon characters, which the build doesn't like. By re-splitting the string by the semicolons I let the engine rejoin them however it pleases, and that plays nicely with MSBuild.

This PR is sent for verification in xBuild. If that works, then it should be ok to merge.