[intents] Enable API using NSMeasurement<UnitType> (#568)
Due to how bindings are defined, e.g. using an [BaseType] attribute for
hierarchy, we cannot restrict the generic types in them. IOW

	interface NSMeasurement<UnitType> where UnitType : NSUnit {

fails when using like

		NSMeasurement<NSUnitTemperature> Temperature { get; }

because `NSUnitTemperature` is not a subclass of `UnitType` in the
bindings and thus this must become something like

	interface NSMeasurement<UnitType> {

to compile (bindings). From there a partial class (for the product
assembly) can be used to add the _correct_ constraint...

Sadly does not work as the generator will default to NSObject when no
constraint is given :(

That will be fixed, in the generator, in a separate commit/PR.