[generator] Fix some CWL warnings that generator emits
When using WrapAttribute on methods generator still tries to create
signatures to interop with ObjC which is incorrect since the pruppose
of Wrap attribute is to create nicer signatures on top of an "ugly"
API so user gets a delightful coding experience.

When generator tries to create a signature for a decorated method with
WrapAttribute to interop with objc and the signature contains a NET Type
(i.e System.Type) it throws a BindingException but it was swallowed
by a try/catch that only printed a not so useful description of the
the problem. Now it throws the actual BindingException when it really
does not know how to create a signature to interop with ObjC on a
*not* decorated method with WrapAttribute.