[Foundation] Fix NSMeasurement generic type inheritance
Our temp.dll won't compile because
NSMeasurement<NSUnitTemperature> temperature is used in intents and
NSUnitTemperature does not conform (inherit) yet from its parent class
NSDimension which inherits from NSUnit. But at this stage the generator
has not run yet and the temp.dll cannot be compiled because the above
informations is expressed in attributes.

So wee get a compiler error:

 The type `Foundation.NSUnitTemperature' cannot be used as
type parameter `UnitType' in the generic type or method
`Foundation.NSMeasurement<UnitType>'.
There is no implicit reference conversion from
`Foundation.NSUnitTemperature' to `Foundation.NSUnit'

Which is totally correct because that info is not available yet. The
fix is simple we just let interface NSUnitTemperature : NSUnit so our
intermediate temp.dll builds and let the generator do its thing, this
will end up adding the Symbol property to NSUnitTemperature which is
correct (already double checked with an ObjC sample).