[foundation] Fix generic constraint for NSMeasurement
The generator does not currently offer flexibility for constraints.

That's made harder since binding files uses attributes, [BaseType], to
define the type hierarchy.

Here we cheat as we can add, non-decorated, interfaces in the bindings
that won't generate anything - they are only to please the compiler or,
in this case, it let us fool the generator into producing the code we
need (because it does not consider the namespace)

Based on Alex elegant attempt to fix this in:
https://github.com/xamarin/xamarin-macios/pull/780

Generated code diff:
https://gist.github.com/spouliot/eeef3491a6e08d9344a19e9bbc632848