[ReplayKit] Fix RPBroadcastConfiguration.VideoCompressionProperties to have the right type and provide a strongly typed version as well.
The video compression property is defined as this in the headers:

    NSDictionary <NSString *, NSObject <NSCoding, NSSecureCoding> *>

so I'm not sure how it ended up being bound as NSDictionary<NSNumber,
NSSecureCoding> (maybe something from an earlier beta?).

I'm also adding a strongly typed version of the dictionary, since the header
documentation says "AVVideoCompressionPropertiesKey in
<AVFoundation/AVVideoSettings.h> for available properties.".

The manual binding is required because the generator doesn't like generic weak
dictionaries (it generates invalid code).