[Foundation] Fix 401s in the NsUrlHandler.
The NSUrlSessionHandler is not dealing with 40s correctly, unfortunatly
we are modifying the Authorization header which is a reserved header as
per apple documentation. We need to hide this issue by dealing with the
challenge in the handler.

The fix checks if it is the very first failure of the challenge and we
do have the header, in that case, reject the request and continue with
the handler execution.

Fixes bug https://bugzilla.xamarin.com/show_bug.cgi?id=42936