[mtouch] Replace FileInfo with calls to File
FileInfo requires a allocation (small) but will call MonoIO::GetFileStat
in its refresh method. This cache a lot of things, that might not be
needed, and it takes time...

E.g. 916 ms our of my 30334 ms build was in FileInfo

That's 3%, not huge overall, but a lot just to query stuff we won't need.
In fact FileInfo.Exists will eventually be calling File.Exists anyway.

Note: File does not have any method to return Size, so those are still
calling FileInfo